Move package from `redux` to `redux.api`.
