+++ api/src/main/java/redux/api/Store.java
@@ -1,4 +1,4 @@
-package redux;
+package redux.api;
 
 /**
  * Coordinates actions and [Reducers][Reducer]. Store has the following responsibilities:
